+++ src/components/screens/login/Login.jsx
@@ -1,5 +1,5 @@
 import React, { useState } from "react";
-import "../css/Login.css";
+import "../../../styles/Login.css";
 
 export default function Login({ setToken }) {
   const [username, setUserName] = useState();
